requirements.txt
fastapi==0.115.4
uvicorn==0.32.0
jinja2==3.1.4
python-dotenv
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
itsdangerous==2.2.0